:bug: Fixed a few bugs
Fixed that the y position of the palette window would always reset on refresh. Additionally fixed a bug that got introduced with the context menu of the individual colors in the palette that would result in mass deletion of colors. Also fixed text being highlighted in the picker lens when using the quick selection method by holding shift